Ensure test selector stripping works for inline template compilation
Prior to this, the `registry.add('htmlbars-ast-plugin', ...)` invocation in `index.js` runs _after_ all of the `htmlbars-ast-plugins` are read and registered for inline JS/TS compilation. It is still before stand alone templates are ran (and that is why it works for `*.hbs` files that are not colocation).

This ensures that `ember-test-selectors` runs _before_ `ember-cli-htmlbars` / `ember-cli-htmlbars-inline-precompile (so that we can ensure that `ember-test-selectors` included runs, and therefore the `htmlbars-ast-plugin` is registered before it is needed for inline template compilation).